requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
